Feat: restrict BoH chaos

+++ code/game/objects/items/weapons/storage/backpack.dm
@@ -67,17 +67,25 @@
 		var/response = alert(user, "This creates a singularity, destroying you and much of the station. Are you SURE?","IMMINENT DEATH!", "Yes", "No")
 		if(response == "Yes")
 			user.visible_message("<span class='warning'>[user] grins as [user.p_they()] begin[user.p_s()] to put a Bag of Holding into a Bag of Holding!</span>", "<span class='warning'>You begin to put the Bag of Holding into the Bag of Holding!</span>")
-			if(do_after(user, 30, target=src))
-				investigate_log("has become a singularity. Caused by [user.key]","singulo")
-				user.visible_message("<span class='warning'>[user] erupts in evil laughter as [user.p_they()] put[user.p_s()] the Bag of Holding into another Bag of Holding!</span>", "<span class='warning'>You can't help but laugh wildly as you put the Bag of Holding into another Bag of Holding, complete darkness surrounding you.</span>","<span class='warning'> You hear the sound of scientific evil brewing! </span>")
-				qdel(W)
-				var/obj/singularity/singulo = new /obj/singularity(get_turf(user))
-				singulo.energy = 300 //To give it a small boost
-				message_admins("[key_name_admin(user)] detonated a bag of holding <A HREF='?_src_=holder;adminplayerobservecoodjump=1;X=[x];Y=[y];Z=[z]'>JMP</a>)")
-				log_game("[key_name(user)] detonated a bag of holding")
-				qdel(src)
+			var/list/play_records = params2list(user.client.prefs.exp)
+			var/livingtime = text2num(play_records[EXP_TYPE_LIVING])
+			if (user.mind.special_role || livingtime > 9000)
+				if(do_after(user, 30, target=src))
+					investigate_log("has become a singularity. Caused by [user.key]","singulo")
+					user.visible_message("<span class='warning'>[user] erupts in evil laughter as [user.p_they()] put[user.p_s()] the Bag of Holding into another Bag of Holding!</span>", "<span class='warning'>You can't help but laugh wildly as you put the Bag of Holding into another Bag of Holding, complete darkness surrounding you.</span>","<span class='warning'> You hear the sound of scientific evil brewing! </span>")
+					qdel(W)
+					var/obj/singularity/singulo = new /obj/singularity(get_turf(user))
+					singulo.energy = 300 //To give it a small boost
+					message_admins("[key_name_admin(user)] detonated a bag of holding <A HREF='?_src_=holder;adminplayerobservecoodjump=1;X=[x];Y=[y];Z=[z]'>JMP</a>)")
+					log_game("[key_name(user)] detonated a bag of holding")
+					qdel(src)
+				else
+					user.visible_message("After careful consideration, [user] has decided that putting a Bag of Holding inside another Bag of Holding would not yield the ideal outcome.","You come to the realization that this might not be the greatest idea.")
 			else
 				user.visible_message("After careful consideration, [user] has decided that putting a Bag of Holding inside another Bag of Holding would not yield the ideal outcome.","You come to the realization that this might not be the greatest idea.")
+				investigate_log("could potentially become a singularity (feature disabled for non-special roles). Caused by [user.key]","singulo")
+				message_admins("[key_name_admin(user)] tried to detonate a bag of holding (feature disabled for non-special roles) <A HREF='?_src_=holder;adminplayerobservecoodjump=1;X=[x];Y=[y];Z=[z]'>JMP</a>)")
+				log_game("[key_name(user)] tried to detonate a bag of holding (feature disabled for non-special roles)")
 	else
 		. = ..()
